Refactor Codebase remove DRY issue
